Node recovery

+++ src/libPersistence/Retriever.cpp
@@ -36,7 +36,7 @@ namespace filesys = boost::filesystem;
 
 Retriever::Retriever(Mediator& mediator) : m_mediator(mediator) {}
 
-void Retriever::RetrieveDSBlocks(bool& result) {
+void Retriever::RetrieveDSBlocks(bool& result, const bool& wakeupForUpgrade) {
   LOG_MARKER();
 
   std::list<DSBlockSharedPtr> blocks;
@@ -75,23 +75,34 @@ void Retriever::RetrieveDSBlocks(bool& result) {
   }
 
   if (isDSIncompleted[0] == '1') {
-    LOG_GENERAL(INFO, "Has incompleted DS Block");
-    if (BlockStorage::GetBlockStorage().DeleteDSBlock(blocks.size() - 1)) {
-      BlockStorage::GetBlockStorage().PutMetadata(MetaType::DSINCOMPLETED,
-                                                  {'0'});
+    /// Removing incompleted DS for upgrading protocol
+    /// Keeping incompleted DS for node recovery
+    if (wakeupForUpgrade) {
+      LOG_GENERAL(INFO, "Has incompleted DS Block, remove it");
+      if (BlockStorage::GetBlockStorage().DeleteDSBlock(blocks.size() - 1)) {
+        BlockStorage::GetBlockStorage().PutMetadata(MetaType::DSINCOMPLETED,
+                                                    {'0'});
+      }
+      blocks.pop_back();
+    } else {
+      LOG_GENERAL(INFO, "Has incompleted DS Block, keep it");
     }
-    blocks.pop_back();
-    hasIncompletedDS = true;
   }
+
+  m_mediator.m_blocklinkchain.Reset();
+  uint64_t index = 0;
 
   for (const auto& block : blocks) {
     m_mediator.m_dsBlockChain.AddBlock(*block);
+    m_mediator.m_blocklinkchain.AddBlockLink(
+        index++, block->GetHeader().GetBlockNum(), BlockType::DS,
+        block->GetBlockHash());
   }
 
   result = true;
 }
 
-void Retriever::RetrieveTxBlocks(bool& result) {
+void Retriever::RetrieveTxBlocks(bool& result, const bool& wakeupForUpgrade) {
   LOG_MARKER();
   std::list<TxBlockSharedPtr> blocks;
   if (!BlockStorage::GetBlockStorage().GetAllTxBlocks(blocks)) {
@@ -104,12 +115,18 @@ void Retriever::RetrieveTxBlocks(bool& result) {
     return a->GetHeader().GetBlockNum() < b->GetHeader().GetBlockNum();
   });
 
-  // truncate the extra final blocks at last
-  int totalSize = blocks.size();
-  int extra_txblocks = totalSize % NUM_FINAL_BLOCK_PER_POW;
-  for (int i = 0; i < extra_txblocks; ++i) {
-    BlockStorage::GetBlockStorage().DeleteTxBlock(totalSize - 1 - i);
-    blocks.pop_back();
+  unsigned int totalSize = blocks.size();
+  unsigned int extra_txblocks = totalSize % NUM_FINAL_BLOCK_PER_POW;
+
+  if (wakeupForUpgrade ||
+      (blocks.back()->GetHeader().GetBlockNum() + NUM_VACUOUS_EPOCHS) %
+              NUM_FINAL_BLOCK_PER_POW ==
+          0) {
+    // truncate the extra final blocks at last
+    for (unsigned int i = 0; i < extra_txblocks; ++i) {
+      BlockStorage::GetBlockStorage().DeleteTxBlock(totalSize - 1 - i);
+      blocks.pop_back();
+    }
   }
 
   for (const auto& block : blocks) {
@@ -117,6 +134,37 @@ void Retriever::RetrieveTxBlocks(bool& result) {
   }
 
   result = true;
+
+  /// If recovery mode with vacuous epoch or less than 1 DS epoch, apply re-join
+  /// process instead of node recovery
+  if (!wakeupForUpgrade &&
+      (blocks.back()->GetHeader().GetBlockNum() < NUM_FINAL_BLOCK_PER_POW ||
+       (blocks.back()->GetHeader().GetBlockNum() + NUM_VACUOUS_EPOCHS) %
+               NUM_FINAL_BLOCK_PER_POW ==
+           0)) {
+    result = false;
+    LOG_GENERAL(INFO,
+                "Node recovery with vacuous epoch or too early, apply "
+                "re-join process instead");
+    return;
+  }
+
+  /// Retrieve final block state delta from last DS epoch to
+  /// current TX epoch
+  for (const auto& block : blocks) {
+    if (block->GetHeader().GetBlockNum() >= totalSize - extra_txblocks) {
+      std::vector<unsigned char> stateDelta;
+      BlockStorage::GetBlockStorage().GetStateDelta(
+          block->GetHeader().GetBlockNum(), stateDelta);
+
+      if (AccountStore::GetInstance().DeserializeDelta(stateDelta, 0) != 0) {
+        LOG_GENERAL(WARNING,
+                    "AccountStore::GetInstance().DeserializeDelta failed");
+        result = false;
+        return;
+      }
+    }
+  }
 }
 
 bool Retriever::CleanExtraTxBodies() {
